Move tests to specs.
