+++ spec/lib/compatibility_stamping_test.rb
