+++ spec/lib/stamping_test.rb
