+++ spec/lib/userstamp_controller_test.rb
