+++ spec/lib/userstamp_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+RSpec.describe ActiveRecord::Userstamp do
+  it 'has a VERSION' do
+    expect(ActiveRecord::Userstamp::VERSION).to match(/^\d+\.\d+\.\d+$/)
+  end
+end
