+++ spec/support/models/foo.rb
