+++ spec/support/routes.rb
@@ -1,0 +1,3 @@
+ActionController::Routing::Routes.draw do |map|
+  map.connect ':controller/:action/:id'
+end
